auctionweb/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
pymysql>=1.1.0
pydantic>=2.0.0
python-multipart>=0.0.6
openpyxl>=3.1.0
tzdata>=2023.3
python-json-logger>=2.0.7
python-dotenv>=1.0.0
Babel


# python -m uvicorn app.main:app --reload
# python -m babel.messages.frontend compile -d app/translations
# docker-compose up -d --build  # Build và chạy lại
# docker-compose build       # Build lại image
# docker-compose restart   # Khởi động lại container
# docker-compose start      # Khởi động container